matcher/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
pypdf2
pymupdf
pdf2image
pytesseract
requests
jinja2
markdown
redis
rq
psycopg2-binary
sqlalchemy
alembic
google-generativeai
openai
